moved files
Moved files related to Get Started tab into separate directory.
