+++ src/components/app-nav-menu/getting-started/__tests__/GetStartedPage.spec.tsx
@@ -15,7 +15,7 @@ import { createHashHistory } from 'history';
 import { render, screen } from '@testing-library/react';
 import { GetStartedPage } from '../GetStartedPage';
 
-jest.mock('../SamplesListTab', () => {
+jest.mock('../get-started-tab/SamplesListTab', () => {
   return function DummyTab(): React.ReactElement {
     return <span>Samples List Tab Content</span>;
   };
