+++ src/components/app-nav-menu/getting-started/__tests__/SamplesListGallery.spec.tsx
@@ -16,7 +16,7 @@ import thunk from 'redux-thunk';
 import createMockStore from 'redux-mock-store';
 import { render, screen, RenderResult, fireEvent } from '@testing-library/react';
 import mockAxios from 'axios';
-import SamplesListGallery from '../SamplesListGallery';
+import SamplesListGallery from '../get-started-tab/SamplesListGallery';
 import { Provider } from 'react-redux';
 import { AppState } from '../../../../store';
 import mockMetadata from './devfileMetadata.json';
