+++ src/components/app-nav-menu/getting-started/__tests__/TemporaryStorageSwitch.spec.tsx
