+++ src/components/app-nav-menu/getting-started/get-started-tab/SampleCard.styl
