+++ src/components/app-nav-menu/getting-started/get-started-tab/SamplesListGallery.tsx
@@ -22,9 +22,9 @@ import {
   Gallery,
   Title,
 } from '@patternfly/react-core';
-import { AppState } from '../../../store';
-import * as DevfileFiltersStore from '../../../store/DevfileFilters';
-import * as DevfileRegistriesStore from '../../../store/DevfileRegistries';
+import { AppState } from '../../../../store';
+import * as DevfileFiltersStore from '../../../../store/DevfileFilters';
+import * as DevfileRegistriesStore from '../../../../store/DevfileRegistries';
 import { SampleCard } from './SampleCard';
 import { SearchIcon } from '@patternfly/react-icons';
